% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campus_reults.R
\name{campus_results}
\alias{campus_results}
\title{An internal function used to generate the QEP master report.}
\usage{
campus_results(data, location)
}
\arguments{
\item{data}{The internal master grade dataframe for the report.}

\item{location}{The campus for which the plot is generated.}
}
\description{
An internal function used to generate the QEP master report.
}
