% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnlgrd_clean.R
\name{fnlgrd_clean}
\alias{fnlgrd_clean}
\title{Clean Final Grades for the QEP.}
\usage{
fnlgrd_clean(data, loc)
}
\arguments{
\item{data}{The dataset to clean.}

\item{loc}{The dataframe that contains Keiser locations and starting terms.}
}
\value{
A dataset that has: (1) abbreviated campus names, (2) ordered terms, (3) separate variables
  for term, semester, and year, (3) no WNAs, (4) class size, (5) attempt number, and
  (6) defines success.
}
\description{
Clean Final Grades for the QEP.
}
