% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keiser_terms.R
\name{keiser_terms}
\alias{keiser_terms}
\title{Generate a list of KU terms.}
\usage{
keiser_terms(start, stop)
}
\arguments{
\item{start}{The starting term entered as text (in quotes), two-digit year, semester character, and term character.}

\item{stop}{The stopping term entered as text (in quotes), two-digit year, semester character, and term character.}
}
\value{
A sequence of ordered term codes.
}
\description{
Generate a list of KU terms.
}
