% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{locations}
\alias{locations}
\title{Keiser QEP Locations}
\format{A data frame with 8 rows and 3 columns.}
\description{
An updated list of full Keiser campus names, and associated campus abbreviations and QEP starts.
}
\details{
\itemize{
\item full Full Keiser University location name.
\item abbreviation Abbreviated Keiser location.
\item start_term Term code when that campus began the QEP.
}
}
