% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_campus_results.R
\name{plot_campus_results}
\alias{plot_campus_results}
\title{An internal function used to generate the QEP master report.}
\usage{
plot_campus_results(data, location)
}
\arguments{
\item{data}{The internal master grade dataframe for the report.}

\item{location}{The campus for which the plot is generated.}
}
\value{
campus_plot A ggplot object.
}
\description{
An internal function used to generate the QEP master report.
}
